add lifetime parameter to the definition of `Endpoint`
In the modified definition of `Endpoint`, the lifetime of the apply()'s
receiver is added to the trait bound of the associated type `Future`.
It allows `Self::Future` to have some references to the inner fields in
`Self`, that is, it is possible to delete unnecessary cloning at calling
`apply()`.

NOTE:

In the future, such treatments should be written using generic
associated types as follows:

```rust
trait Endpoint {
    type Out;
    type Fut<'a>: TryFuture + 'a;

    fn apply<'a>(&'a self) -> Result<Self::Fut<'a>, Error>;
}
```